requirements.txt
django
Pillow
python-dotenv
psycopg2-binary